requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.4.1
annotated-types==0.7.0
anyio==4.8.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
constantly==15.1.0
contourpy==1.3.1
cryptography==43.0.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
dnspython==2.7.0
et_xmlfile==2.0.0
executing==2.0.1
faiss-cpu==1.10.0
filelock==3.18.0
filetype==1.2.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.1.24
fonttools==4.55.3
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.3.2
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.67.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h5py==3.12.1
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
hyperlink==21.0.0
idna==3.4
imutils==0.5.4
incremental==22.10.0
ipykernel==6.29.5
ipynb-py-convert==0.4.6
ipython==8.26.0
ipywidgets==8.1.5
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kaggle==1.6.17
keras==3.8.0
kiwisolver==1.4.7
langchain==0.3.19
langchain-community==0.3.17
langchain-core==0.3.35
langchain-google-genai==2.0.9
langchain-text-splitters==0.3.6
langsmith==0.3.8
libclang==18.1.1
logzero==1.7.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.0.2
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
peewee==3.17.8
pillow==11.0.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.0.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.11.1
pyotp==2.8.0
pyparsing==3.1.0
PyPDF2==3.0.1
pyreadstat==1.2.8
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
pywin32==306
PyYAML==6.0.2
pyzmq==26.0.3
rdflib==6.3.2
rdflib-jsonld==0.6.2
reportlab==4.3.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.8.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
setuptools==75.8.0
simplejson==3.19.1
six==1.16.0
smartapi-python==1.4.8
sniffio==1.3.1
sns==0.1
soupsieve==2.6
SQLAlchemy==2.0.38
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
torch==2.6.0
torchvision==0.21.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
Twisted==22.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ucimlrepo==0.0.7
uritemplate==4.1.1
urllib3==2.0.3
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.2
yarl==1.18.3
yfinance==0.2.50
zope.interface==6.0
zstandard==0.23.0
